refactor: moved caching functions to separate service
