+++ services/utility.service.js
@@ -1,5 +1,3 @@
-import * as fs from 'fs';
-import * as path from 'path';
 import { directionValues } from '../models/posts-request/directionValues.model.js';
 
 export const UtilityService = {};
@@ -34,28 +32,4 @@ UtilityService.sortPosts = (arr, sortBy, direction) => {
         postA[sortBy] - postB[sortBy] : // Descending
       -1*(postA[sortBy] - postB[sortBy]); // Ascending
     });
-}
-
-UtilityService.cachePosts = (posts, tag) => {
-  // if(UtilityService.matchCacheFiles(tag) === -1) {
-  fs.writeFileSync(path.resolve(`cache/${tag}.json`), JSON.stringify(posts));
-  // }
-}
-
-UtilityService.getCachedPosts = (tag) => {
-  const files = fs.readdirSync(path.resolve('cache'));
-  let fileName = UtilityService.matchCacheFiles(tag);
-  if(!!fileName) {
-    const cachedFile = fs.readFileSync(path.resolve(`cache/${fileName}`));
-    return JSON.parse(cachedFile);
-  }
-   return undefined;
-}
-
-UtilityService.matchCacheFiles = (tag) => {
-  //todo: find a way to make this non-blocking
-  const files = fs.readdirSync(path.resolve('cache'));
-
-  const regex = new RegExp(tag + '.', 'g');
-  return files.find((file) => file.match(regex));
 }
